requirements.txt
pytest~=8.3.3
allure-pytest==2.13.5
allure-python-commons==2.13.5
PyYAML==6.0.1
selenium==4.24.0
webdriver-manager==4.0.2